# Optimize SplFixedArray::fromArray() for packed arrays